KW-538/refactor: amdin-> entrance 디렉토리, 프로필 명 수정
