+++ src/main/java/com/doubleo/entrancegateway/filter/JwtAuthenticationFilter.java
@@ -1,7 +1,7 @@
-package com.doubleo.admingateway.filter;
+package com.doubleo.entrancegateway.filter;
 
-import com.doubleo.admingateway.infra.config.jwt.JwtProperties;
-import com.doubleo.admingateway.infra.config.redis.BlackListTokenService;
+import com.doubleo.entrancegateway.infra.config.jwt.JwtProperties;
+import com.doubleo.entrancegateway.infra.config.redis.BlackListTokenService;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.JwtException;
 import io.jsonwebtoken.Jwts;
