+++ src/main/java/com/doubleo/entrancegateway/infra/config/properties/PropertiesConfig.java
@@ -1,8 +1,8 @@
-package com.doubleo.admingateway.infra.config.properties;
+package com.doubleo.entrancegateway.infra.config.properties;
 
-import com.doubleo.admingateway.infra.config.jwt.JwtProperties;
-import com.doubleo.admingateway.infra.config.redis.RedisProperties;
-import com.doubleo.admingateway.infra.config.route.GatewayPathProperties;
+import com.doubleo.entrancegateway.infra.config.jwt.JwtProperties;
+import com.doubleo.entrancegateway.infra.config.redis.RedisProperties;
+import com.doubleo.entrancegateway.infra.config.route.GatewayPathProperties;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
