+++ src/main/resources/application.yml
@@ -1,6 +1,6 @@
 spring:
   application:
-    name: admin-gateway
+    name: entrance-gateway
 
   config:
     import: optional:configserver:${CONFIG_SERVER_URL}
