+++ src/test/java/com/doubleo/entrancegateway/EntranceGatewayApplicationTests.java
@@ -1,4 +1,4 @@
-package com.doubleo.admingateway;
+package com.doubleo.entrancegateway;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -6,7 +6,7 @@ import org.springframework.test.context.ActiveProfiles;
 
 @SpringBootTest
 @ActiveProfiles({"test", "redis", "security", "eureka", "swagger", "routes", "actuator"})
-class AdminGatewayApplicationTests {
+class EntranceGatewayApplicationTests {
 
     @Test
     void contextLoads() {}
